+++ Kinder/Database/UsersContext.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 
-namespace Kinder
+namespace Kinder.Database
 {
     class UsersContext : DbContext
     {
